requirements.txt
python-dotenv==0.15.0
pytelegrambotapi==3.7.6
wheel==0.36.2
Proxy-List-Scrapper==0.2.1